Schedule calls & auto call disconnect: temporary fixes. Replace all the usbfd & adspfd identifiers for actual ints Bump version to 0.5.9

+++ recipes-modem/openqti/files/inc/call.h
@@ -269,14 +269,14 @@ void set_pending_call_flag(bool en);
 uint8_t get_call_pending();
 void set_call_simulation_mode(bool en);
 uint8_t get_call_simulation_mode();
-void start_simulated_call(uint8_t usbfd);
+void start_simulated_call(int usbfd);
 
-void send_call_request_response(uint8_t usbfd, uint16_t transaction_id);
-uint8_t send_voice_call_status_event(uint8_t usbfd, uint16_t transaction_id,
+void send_call_request_response(int usbfd, uint16_t transaction_id);
+uint8_t send_voice_call_status_event(int usbfd, uint16_t transaction_id,
                                      uint8_t call_direction,
                                      uint8_t call_state);
-void send_dummy_call_established(uint8_t usbfd, uint16_t transaction_id);
+void send_dummy_call_established(int usbfd, uint16_t transaction_id);
 uint8_t call_service_handler(uint8_t source, void *bytes, size_t len,
-                             uint16_t msgid, uint8_t adspfd, uint8_t usbfd);
+                             uint16_t msgid, int adspfd, int usbfd);
 
 #endif
