+++ recipes-modem/openqti/files/inc/sms.h
@@ -276,14 +276,14 @@ uint8_t ack_message_notification(int fd, uint8_t pending_message_num);
 uint8_t inject_message(uint8_t message_id);
 uint8_t do_inject_notification(int fd);
 
-uint8_t intercept_and_parse(void *bytes, size_t len, uint8_t hostfd, uint8_t adspfd);
+uint8_t intercept_and_parse(void *bytes, size_t len, int hostfd, int adspfd);
 
 
 int process_message_queue(int fd);
 void add_message_to_queue(uint8_t *message, size_t len);
 void notify_wms_event(uint8_t *bytes,size_t len, int fd);
-int check_wms_message(void *bytes, size_t len, uint8_t adspfd, uint8_t usbfd);
-int check_wms_indication_message(void *bytes, size_t len, uint8_t adspfd,
-                                 uint8_t usbfd);
+int check_wms_message(void *bytes, size_t len, int adspfd, int usbfd);
+int check_wms_indication_message(void *bytes, size_t len, int adspfd,
+                                 int usbfd);
 
 #endif
